requirements.txt
Django
djangorestframework
django-cors-headers
pymongo
pytest
pytest-django
